Progress towards project1 revision
Added an enum to represent each instruction type. Changed instruction_id to an enum ID. Changed label_id to a String and changed the name to label. Changed the branch_label to String.
Made changes to functions affected by the type and name changes.

Added methods for creating each instruction based on type in the factory.